Incremental commit
In the process of plugging all of the api together and figuring out how to properly winnow stops.

Some major bugs, and large optimization issues—when I'm done with this I want to get it down to a single streaming api call, instead of one per tracked stop. Some of this code is deeply ugly. It's coming along, though.